# Adds a few performance tracking scenarios 🎰