--- a/test/fixtures/index.ts
+++ b/test/fixtures/index.ts
@@ -1,0 +1,2 @@
+export * from "./fixture-data"
+export * from "./fixture-models"