--- a/test/perf.ts
+++ b/test/perf.ts
@@ -1,5 +1,6 @@
 import { test } from "ava"
 import { smallScenario, mediumScenario, largeScenario } from "./perf/scenarios"
+import { start } from "./perf/timer"
 
 // TODO: Not sure how this should work. This feels super fragile.
 const TOO_SLOW_MS = 1000
@@ -17,4 +18,18 @@ test.serial("performs well on large scenario", t => {
     t.true(largeScenario(10, 10, 0).elapsed < TOO_SLOW_MS)
     t.true(largeScenario(10, 0, 10).elapsed < TOO_SLOW_MS)
     t.true(largeScenario(10, 10, 10).elapsed < TOO_SLOW_MS)
+})
+
+test.cb("timer", t => {
+    const go = start()
+    setTimeout(function() {
+        const lap = go(true)
+        setTimeout(function() {
+            const done = go()
+            t.not(lap, 0)
+            t.not(done, 0)
+            t.not(lap, done)
+            t.end()
+        }, 2)
+    }, 2)
 })