--- a/test/perf/timer.ts
+++ b/test/perf/timer.ts
@@ -16,9 +16,9 @@
  */
 export const start = () => {
     const started = process.hrtime()
-    let last = started
+    let last: [number, number] = [started[0], started[1]]
     return (lapTime: boolean = false): number => {
-        last = process.hrtime(lapTime ? last : started)
-        return Math.round((last[0] * 1e9 + last[1]) / 1e6)
+        const final = process.hrtime(lapTime ? last : started)
+        return Math.round((final[0] * 1e9 + final[1]) / 1e6)
     }
 }