I'm pretty sure this is the opposite of "science"... but... here we go.

### Observations

* having many attributes does seem to slow things down
* haven't really zeroed in on why yet
* nested children really tank perf (seeming to hint at a fixed cost per-node)

Any other scenarios you'd like to see?

### Usage

```
npm run speedtest
```

<details><summary>Incoming Wall of Text</summary>

```
-----------
Small Model
-----------
      31ms |      x 100 |  0.3ms avg
     141ms |    x 1,000 |  0.1ms avg
   1,314ms |   x 10,000 |  0.1ms avg
     128ms |    x 1,000 |  0.1ms avg
      32ms |      x 100 |  0.3ms avg

------------
Medium Model
------------
      38ms |      x 100 |  0.4ms avg
     262ms |    x 1,000 |  0.3ms avg
   3,320ms |   x 10,000 |  0.3ms avg
     289ms |    x 1,000 |  0.3ms avg
      31ms |      x 100 |  0.3ms avg

------------------------
Large Model - 0 children
------------------------
     164ms |      x 100 |  1.6ms avg
   1,655ms |    x 1,000 |  1.7ms avg
     166ms |      x 100 |  1.7ms avg

-------------------------------------------
Large Model - 10 small & 10 medium children
-------------------------------------------
     278ms |       x 50 |  5.6ms avg
   1,546ms |      x 250 |  6.2ms avg
     268ms |       x 50 |  5.4ms avg

-------------------------------------------
Large Model - 100 small & 0 medium children
-------------------------------------------
     882ms |       x 50 | 17.6ms avg
   4,484ms |      x 250 | 17.9ms avg
     841ms |       x 50 | 16.8ms avg

-------------------------------------------
Large Model - 0 small & 100 medium children
-------------------------------------------
   1,734ms |       x 50 | 34.7ms avg
  11,082ms |      x 250 | 44.3ms avg
   1,778ms |       x 50 | 35.6ms avg
```

</details>

### Tests
I added `test/perf.ts`, but i'm not convinced it's a good idea as it makes the tests a bit brittle. 

TBH: I don't have much experience in this area (as you can probably see!) 😊 

### Refs

Follows up on #284.
Ghetto pre-cursor to #254.