@mattiamanzati 2017-08-04
I think this report is great and will help a lot keeping track of performance! :)
Maybe we want to write the test to FS and run as precommit hook? What do you think?

cc @mweststrate