@skellock 2017-08-04
Maybe save it as `json` with a key that is linked to a type of computer somehow so we compare 🍎  to 🍎 ?

```ts
const os = require('os')
`${process.platform} ${os.arch()} ${os.cpus()[0].model} ${os.cpus().length} cores`

// darwin x64 Intel(R) Core(TM) i7-4790 CPU @ 3.60GHz 8 cores
```