@mweststrate 2017-08-05
superb work @skellock! I will start using this as base for optimizations as soon as #286 is done!